Рефакторинг компонента таблицы задач
